+++ src/main/java/com/blackbuild/groovy/configdsl/transform/DSL.java
@@ -25,6 +25,7 @@ package com.blackbuild.groovy.configdsl.transform;
 
 import com.blackbuild.groovy.configdsl.transform.ast.DSLASTTransformation;
 import org.codehaus.groovy.transform.GroovyASTTransformationClass;
+import org.codehaus.groovy.transform.StaticTypesTransformation;
 
 import java.lang.annotation.*;
 
@@ -418,7 +419,7 @@ import java.lang.annotation.*;
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.CLASS)
 @Inherited // This is currently not used, see https://issues.apache.org/jira/browse/GROOVY-6765
-@GroovyASTTransformationClass(classes={DSLASTTransformation.class})
+@GroovyASTTransformationClass(classes={DSLASTTransformation.class, StaticTypesTransformation.class})
 @Documented
 public @interface DSL {
 }